# Use 16-bit depth texture for screen-space reflections