--- a/servers/rendering/renderer_rd/shaders/screen_space_reflection.glsl
+++ b/servers/rendering/renderer_rd/shaders/screen_space_reflection.glsl
@@ -7,7 +7,7 @@
 layout(local_size_x = 8, local_size_y = 8, local_size_z = 1) in;
 
 layout(rgba16f, set = 0, binding = 0) uniform restrict readonly image2D source_diffuse;
-layout(r32f, set = 0, binding = 1) uniform restrict readonly image2D source_depth;
+layout(r16f, set = 0, binding = 1) uniform restrict readonly image2D source_depth;
 layout(rgba16f, set = 1, binding = 0) uniform restrict writeonly image2D ssr_image;
 #ifdef MODE_ROUGH
 layout(r8, set = 1, binding = 1) uniform restrict writeonly image2D blur_radius_image;