--- a/servers/rendering/renderer_rd/shaders/screen_space_reflection_filter.glsl
+++ b/servers/rendering/renderer_rd/shaders/screen_space_reflection_filter.glsl
@@ -14,7 +14,7 @@ layout(rgba16f, set = 2, binding = 0) uniform restrict writeonly image2D dest_ss
 #ifndef VERTICAL_PASS
 layout(r8, set = 2, binding = 1) uniform restrict writeonly image2D dest_radius;
 #endif
-layout(r32f, set = 3, binding = 0) uniform restrict readonly image2D source_depth;
+layout(r16f, set = 3, binding = 0) uniform restrict readonly image2D source_depth;
 
 layout(push_constant, std430) uniform Params {
 	vec4 proj_info;