This improves performance by using a lower precision texture, similar to what is being done in SSAO and SSIL already.

On a GTX 1080 in 2560×1440, this saves 0.1 ms of frametime for a nearly identical appearance.

## Performance

Average over 10 seconds (two runs each):

| Before | After |
|-|-|
| 312 FPS (3.20 mspf) | 322 FPS (3.10 mspf) |

## Preview

### Before

### After