@fire 2022-02-28
@Calinou Can you provide a false color image difference image.

I've been using reg-cli.

![image](https://user-images.githubusercontent.com/32321/155917567-da808598-0264-4aa8-8948-b8561fd86abc.png)