@clayjohn 2022-02-28
SSR can be very sensitive to changes in depth precision in a way that SSAO and SSIL are not because it captures much more high frequency detail. We should test this on a more complex scene with high-frequency detail to properly evaluate its cost/benefit.