@Calinou 2022-02-28
I've made a more complex test scene: [test_ssr_depth.zip](https://github.com/godotengine/godot/files/8154793/test_ssr_depth.zip)

### Before

![2022-02-28_17 06 01_before](https://user-images.githubusercontent.com/180032/156017104-d238f92d-98a6-485e-8418-50a696afb7d3.png)

### After

![2022-02-28_17 07 49_after](https://user-images.githubusercontent.com/180032/156017098-75d42865-151c-4080-946b-c2f7f44a2b04.png)

### Difference ([dssim](https://github.com/kornelski/dssim))

![ssr_diff-0](https://user-images.githubusercontent.com/180032/156017085-2c77be64-c88b-444a-8993-08fcd7d39e7e.png)

Depending on where you look, the difference is non-negligible in this scene, but there are some caveats to keep in mind:

- When you look at the middle of the scene, the difference is barely noticeable.
- This scene has the number of SSR steps doubled to 128 (from 64).
- This scene only uses non-rough, metallic materials where SSR is the most visible.
- SSR could perhaps have a distance cutoff in the distance to avoid artifacting (like SSAO and SSIL already do). This is also noticeable when using a 32-bit depth texture anyway.