@Calinou 2022-11-11
> What is the status on this PR?

We're not sure on whether it's actually a good idea. Exploring other optimizations such as Hi-Z may be more worthwhile without decreasing quality.